Separated ArtistRepository.kt from ArtistRepositoryImpl.kt and moved ArtistRepository.kt to repository package and ArtistRepositoryImpl.kt to data package
