+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/domain/repository/ArtistRepository.kt
@@ -1,0 +1,10 @@
+package ayds.lisboa.songinfo.moredetails.fulllogic.domain.repository
+
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.entities.Artist
+
+interface ArtistRepository {
+    fun getArtist(artistName: String): Artist
+
+}
+
+
